% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestStatistics.R
\name{AtST_stat}
\alias{AtST_stat}
\title{Statistics Calculation of AtST}
\usage{
AtST_stat(
  data.list,
  an = 0.15 * nrow(data.list$X)/log(nrow(data.list$X)),
  bn = 2.25 * log(log(nrow(data.list$X)))/(log((log(log(ncol(data.list$M)) + 1) + 1)) +
    1)
)
}
\arguments{
\item{data.list}{a list with 3 elements, including X, A, Y}

\item{an}{the tuning parameter in the indicators}

\item{bn}{the tuning parameter in the test statistic}
}
\value{
a list of 3 elements, including test statistics, T statistic of beta, and T statistic of gamma.
}
\description{
This function is to calculate the test statistics with data input.
}
\examples{
data(data_list_example)
AtST_stat(data.list.example)$TestStat
}
