% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDRcontrol.R
\name{SeqHyTest_FDR}
\alias{SeqHyTest_FDR}
\title{Sequential Hypothesis Testing with FDR controlled}
\usage{
SeqHyTest_FDR(p_value.seq, tuning_type, tuning, FDR)
}
\arguments{
\item{p_value.seq}{p-value sequence}

\item{tuning_type}{indicator to be chosen from either either 'DOS' or 'Storey'}

\item{tuning}{tuning parameter for DOS or Storey method}

\item{FDR}{the expected FDR}
}
\value{
list of 2 elements, including the cutoff, and the rejection indicator
}
\description{
This function is to calculate cutoff value, and the corresponding rejection
result, with p-value sequence, tuning parameter selection method, and
controlled FDR input.
}
\examples{
data(data_list_example)
TStat = TestStatistics(data.list.example)
p_value.seq = 1-pchisq(q=TStat,df=1)
SeqHyTest_FDR(p_value.seq,tuning_type='DOS',tuning=1,FDR=0.2)
}
