% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestStatistics.R
\name{TestStatistics}
\alias{TestStatistics}
\title{Statistics Calculation}
\usage{
TestStatistics(data.list)
}
\arguments{
\item{data.list}{a list with 3 elements, including X, A, Y}
}
\value{
test statistics
}
\description{
This function is to calculate the test statistics with data input.
}
\examples{
data(data_list_example)
TestStatistics(data.list.example)
}
